Replace #to_ical with #ical
So far #to_ical was returning valid iCal strings while #ical simply
returned nil. Now #ical is the only method to get the iCal
representation of an order_frequency and it's consistent with #ical=
setter.

However, the API responses still contain the 'to_ical' attribute so as
not to break the client.